Fix name="none" and c:type="void" handling for "alias" nodes
Alias nodes may reference 'untyped' data, and thus implicitly
declare an opaque type. gir2pascal needs to be aware of this,
as the current handling via the CType "void" doesn't work.

We want to add a corresponding 'fuzzy' type and use that when
emitting the Pascal type code in TPascalUnit.HandleAlias().

This is a first version of code that properly handles the new
variant of "alias" nodes. With this, Poppler-0.18.gir and all
of its dependencies get converted without later syntax errors.